Restructure controllers
