+++ package.json
@@ -5,7 +5,9 @@
   "main": "./lib/main.js",
   "scripts": {
     "lint": "eslint . --ext .ts",
-    "test": "nyc ts-mocha --no-deprecation -t 60000 --exit test/**/*.ts",
+    "test": "nyc ts-mocha --no-deprecation -t 60000 --exit test/**/*.ts test/**/**/*.ts",
+    "echo": "echo \"test/**/*.ts\"",
+    "echo2": "echo test/**/*.ts",
     "build": "tsc -p .",
     "build:live": "concurrently \"tsc -w -p .\" \"nodemon ./lib/main.js\"",
     "start": "node ./lib/main.js",
@@ -46,6 +48,8 @@
     "eslint-config-airbnb-typescript-prettier": "^4.2.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
+    "glob": "^8.0.1",
+    "minimatch": "^5.0.1",
     "mocha": "^8.1.3",
     "mongodb-memory-server": "^8.0.2",
     "mongoose-to-swagger": "^1.3.0",
