+++ src/app/controllers/authors.ts
@@ -3,12 +3,12 @@ import mongoose from 'mongoose';
 import restify from 'express-restify-mongoose';
 import * as DocumentTypes from '../models/interfaces';
 import { APIOptions } from '../../config/interfaces';
-import { PaperStats } from '../../types';
+
 const passport = require('passport');
 
 export function initialize(
   model: mongoose.Model<DocumentTypes.Author>,
-  router: express.Router,
+  router: express.Router | express.Application,
   options: APIOptions
 ) {
   // authors endpoint
@@ -55,25 +55,4 @@ export function initialize(
       });
     },
   });
-
-  router.get(
-    `${options.server.baseRoute}/fe/authors/list`,
-    passport.authenticate('jwt', { session: false }),
-    async (req: express.Request, res: express.Response) => {
-      const pattern = req.query.pattern;
-      if (!pattern) {
-        res.status(422).json({
-          message: 'The request is missing the required parameter "pattern".',
-        });
-      } else {
-        try {
-          const authorData = await model.find({ fullname: { $regex: pattern } }, { fullname: 1 });
-          console.log(authorData.length);
-          res.json(authorData);
-        } catch (error: any) {
-          res.status(500).json({ message: error.message });
-        }
-      }
-    }
-  );
 }
