+++ src/types.ts
@@ -1,3 +1,5 @@
+import mongoose from 'mongoose';
+
 interface PaperOverTime {
   year: string;
   papers: number;
@@ -20,4 +22,20 @@ export interface PaperJson {
   title: string;
   venues: string;
   authors: string[];
+}
+
+export interface FilterQuery {
+  yearStart: string;
+  yearEnd: string;
+  author?: string;
+  venue?: string;
+}
+
+export interface FilterMongo {
+  datePublished?: {
+    $gt?: Date;
+    $lt?: Date;
+  };
+  authors?: mongoose.Types.ObjectId;
+  venues?: mongoose.Types.ObjectId;
 }
